Обновление

+++ app/src/main/java/com/example/todolist/TaskDB.java
@@ -22,26 +22,26 @@ public class TaskDB extends SQLiteOpenHelper {
 
     static SQLiteDatabase db;//инициализируем бд
 
-    TaskDB(Context context) {
+    TaskDB(Context context) { //конструктор класса TaskDB
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
-        db = this.getWritableDatabase();
-        Log.d("DB_debug", "Creating DBHandler");
+        db = this.getWritableDatabase(); //для возможности записывать в базу
+        //Log.d("DB_debug", "Creating DBHandler");
 //        onUpgrade(db, 2, 3);
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {//создаем таблицу
-        Log.d("DB_debug", "Creating TABLE");
+        //Log.d("DB_debug", "Creating TABLE");
         String CREATE_TASKS = "CREATE TABLE IF NOT EXISTS " + TABLE_TASKS + "(" + KEY_ID + " INTEGER PRIMARY KEY, " + KEY_NAME + " TEXT," + KEY_DESCRIPTION + " TEXT," + KEY_PRIORITY + " TEXT," + KEY_DEADLINE + " TEXT," + KEY_ISDONE + " BIT);";
-        Log.d("DB_debug", CREATE_TASKS);
+        //Log.d("DB_debug", CREATE_TASKS);
         db.execSQL(CREATE_TASKS);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // Drop older table if exist
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_TASKS);
-        // Create tables again
-        onCreate(db);
+        if (oldVersion != newVersion){
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_TASKS);
+            onCreate(db);
+        }
     }
 }
